save_money/python/api/oauth/requirements.txt
# pip install "uvicorn[standard]"
fastapi==0.74.1
sqlalchemy==1.4.31
pydantic==1.9.0
mysqlclient==2.1.1
uvicorn==0.18.3
PyMySQL==1.0.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.0
pydantic-aioredis==0.7.0
python-multipart==0.0.5